branch/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in
certifi==2024.6.2
    # via requests
charset-normalizer==3.3.2
    # via requests
idna==3.7
    # via requests
python-dotenv==1.0.1
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
urllib3==2.2.1
    # via requests
